Fix bugginess in recursive_list()
Previously, 'find' would crash if given a path with a trailing slash
(including just '/' for the root directory itself), due to python's
os.path.basename() function being a bit weird and returning an empty
string in that case.  To work around this, the isdot() function is now
a bit more flexible and accepts empty strings, since it no longer
indexes into its argument.