Fork persist-mode server process earlier
Previously the server process was only forked off after we finished
doing our work with the fsfile, just before exiting.  We now instead
spin up the server process immediately (as soon as we've opened the
fsfile) and then reopen the fsfile through it, doing all operations
via the server socket.

This addresses a somewhat irritating usability problem wherein if you
unlock the fsfile successfully with persist mode enabled but with some
minor invocation mistake that leads to a subsequent error, you don't
get the server process started and hence have to re-enter the password
to re-unlock the fsfile when you run jemf again with the mistake
corrected.  With this change you get the background process going as
soon as you've entered the password correctly, and it sticks around
even if you've biffed something else so you don't have to re-enter it
when you re-run the command.