Simplify type/argument relationship in path-lookup functions
Previously the 'noexc' argument controlled whether a lookup failure
would throw an exception or return None, which made typing awkward
(the return type was sometimes-Optional, which meant it had to be
treated as Optional, which forced a bunch of uncertainty onto callers
that didn't pass noexc=True).  Now we just unconditionally throw an
exception on lookup failures, and its up to the callers previously
passing noexc=True to catch and handle it as appropriate.