Replace is{dir,file,link}() functions with open-coded isinstance() calls
mypy sadly can't do the interprocedural analysis needed to see that
isdir(x) is essentially equivalent to isinstance(x, Directory), which
necessitated a bunch of ugly redundant type assertions and so forth.
While the dedicated functions were nice in terms of being clear,
succinct, and concise, it's not a huge loss.