mv: ensure directory parent entry is updated on cross-directory renames
This fixes a bug that had been lurking for quite a while.  A directory
that was renamed into a different parent directory would be left with
its '..' entry pointing to its old parent instead of the new one.

Also add a test that checks for the bug (though it's not a great test
because it relies on 'pwd' calling Jemf.get_dirname() and that
traversing the chain of directory parents back to the root and failing
one of them is screwed up).